// src/types.ts
export interface CourseData {
  name: string;
  url: string;
  coursetype: string;
  address: string;
  city: string;
  county: string;
  state: string;
  zip: string;
  country: string;
  phone: string;
  web: string;
  twitter: string;
  facebook: string;
  instagram: string | null;
  rangeballs: string;
  tips: string;
  season: string;
  cost: string;
  proshop: string;
  tees: string;
  balls: string;
  instruction: string;
  locker: string;
  par: string;
  yards: string;
  rating: string;
  slope: string;
  architect: string;
  caddie: string;
  banquet: boolean;
  signaturehole: string;
  opened: string;
  greens: string;
  fairways: string;
  waterhazards: boolean;
  sandbunkers: string;
  holes: string;
  yardagemarkers: string;
  acceptteetimes: boolean;
  earliestcallteetime: string;
  trainingfacilities: string;
  onsitegolfpro: string;
  spikes: boolean;
  guests: string;
  access: string;
  discounts: string;
  rentals: boolean;
  pullcarts: boolean;
  walking: boolean;
  restaurant: string;
  bar: boolean;
  hours: string;
  food: boolean;
  availableproducts: string;
  homes: boolean;
  latitude: string;
  longitude: string;
  description: string;
  scorecard: string;
  image: string;
  region: string;
}